requirements.txt
# Core ML libraries
torch==2.1.0
torchvision==0.16.0
pytorch-lightning==2.1.0
lightning==2.1.0

# Configuration and experiment tracking
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0
wandb==0.16.0
omegaconf==2.3.0

# Model optimization and inference
onnx==1.15.0
onnxruntime==1.16.3
onnxsim==0.4.33

# Data versioning
dvc[s3]==3.30.0
boto3==1.34.0

# API and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Data processing
numpy==1.24.4
pandas==2.1.4
scikit-learn==1.3.2
Pillow==10.1.0

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
tqdm==4.66.1
matplotlib==3.8.2
seaborn==0.13.0
